+++ include/nstd/gl/texture/texture.h
@@ -18,9 +18,15 @@ typedef NSTDAny NSTDGLTextureFormat;
 /// Returns: `NSTDGLSurfaceTexture surface_texture` - The surface's texture.
 NSTDAPI NSTDGLSurfaceTexture nstd_gl_surface_texture_current(const NSTDGLSurface surface);
 
+/// Presents a surface texture to the surface.
+/// NOTE: This function will free `surface_texture`.
+/// Parameters:
+///     `NSTDGLSurfaceTexture *const surface_texture` - The surface texture to present.
+NSTDAPI void nstd_gl_surface_texture_present(NSTDGLSurfaceTexture *const surface_texture);
+
 /// Frees a surface texture.
 /// Parameters:
-///     `NSTDGLSurfaceTexture surface_texture` - The surface texture to free.
+///     `NSTDGLSurfaceTexture *const surface_texture` - The surface texture to free.
 NSTDAPI void nstd_gl_surface_texture_free(NSTDGLSurfaceTexture *const surface_texture);
 
 /// Gets the surface's preferred format.